Set default Redis timeout to 100ms.
As indicated in the algorithm description [0], an implementation should
set the Redis communication timeout to something small compared to the
maximum validity time (TTL). As @antirez stated [1] it should be about
two order of magnitude smaller. Given expected TTL times in the 1 - 10
seconds range (depending on the amount of code to be executed in a
caller's critical section), this should be in the 10 - 100 ms range. Set
to 100ms per default.

[0] http://redis.io/topics/distlock
[1] https://github.com/antirez/redlock-rb/issues/3