Permits initialization with existing Redis objects
This allows the Redlock gem to seamlessly work with redis sentinel, which is supported in redis 3.2+.

It also allows clients to set any other arbitrary options on the Redis connection, e.g. password, driver, and more

Bumped version to 0.1.0 per semantic versioning guidelines since functionality was added in a backwards-compatible manner.

Also changed the set command sent to Redis::Client to use 'NX' and 'PX' instead of `:nx` and `:px`. This matches
the behavior of `Redis#set` which parses `{:nx => true, :px => 123}` into `['NX', 'PX', 123]`. This was needed in order
for the redlock gem to work with the fakeredis gem (an in-memory fake redis server commonly used in CI/Test environments).